refactor: move leaderboard functionality to activities endpoints
